requirements.txt
pyyaml>=6.0
requests>=2.28
beautifulsoup4>=4.12
python-dotenv>=1.0